Move migration and vm tests out of pod_test.go
