+++ pkg/virt-controller/watch/vm_test.go
@@ -14,7 +14,9 @@ import (
 	"k8s.io/client-go/pkg/util/workqueue"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/cache"
-	"kubevirt.io/kubevirt/pkg/api/v1"
+
+	kubev1 "k8s.io/client-go/pkg/api/v1"
+	v1 "kubevirt.io/kubevirt/pkg/api/v1"
 	"kubevirt.io/kubevirt/pkg/kubecli"
 	"kubevirt.io/kubevirt/pkg/logging"
 	"kubevirt.io/kubevirt/pkg/virt-controller/services"
@@ -23,9 +25,12 @@ import (
 var _ = Describe("VM watcher", func() {
 	var server *ghttp.Server
 	var migrationCache cache.Store
-
+	var podCache cache.Store
+	var podDispatch kubecli.ControllerDispatch
 	var vmService services.VMService
 	var restClient *rest.RESTClient
+
+	var vmCache cache.Store
 
 	logging.DefaultLogger().SetIOWriter(GinkgoWriter)
 	var dispatch kubecli.ControllerDispatch
@@ -39,7 +44,7 @@ var _ = Describe("VM watcher", func() {
 		clientSet, _ := kubernetes.NewForConfig(&config)
 		templateService, _ := services.NewTemplateService("kubevirt/virt-launcher")
 		restClient, _ = kubecli.GetRESTClientFromFlags(server.URL(), "")
-
+		vmCache = cache.NewIndexer(cache.DeletionHandlingMetaNamespaceKeyFunc, nil)
 		g.Provide(
 			&inject.Object{Value: restClient},
 			&inject.Object{Value: clientSet},
@@ -51,7 +56,8 @@ var _ = Describe("VM watcher", func() {
 		dispatch = NewVMControllerDispatch(restClient, vmService)
 
 		migrationCache = cache.NewIndexer(cache.DeletionHandlingMetaNamespaceKeyFunc, nil)
-
+		podCache = cache.NewIndexer(cache.DeletionHandlingMetaNamespaceKeyFunc, nil)
+		podDispatch = NewPodControllerDispatch(vmCache, restClient, vmService, clientSet)
 	})
 
 	Context("Creating a VM ", func() {
@@ -109,6 +115,56 @@ var _ = Describe("VM watcher", func() {
 			dispatch.Execute(migrationCache, queue, key)
 
 			Expect(len(server.ReceivedRequests())).To(Equal(3))
+			close(done)
+		}, 10)
+	})
+
+	Context("Running Pod for unscheduled VM given", func() {
+		It("should update the VM with the node of the running Pod", func(done Done) {
+
+			// Create a VM which is being scheduled
+			vm := v1.NewMinimalVM("testvm")
+			vm.Status.Phase = v1.Scheduling
+			vm.ObjectMeta.SetUID(uuid.NewUUID())
+
+			// Add the VM to the cache
+			vmCache.Add(vm)
+
+			// Create a target Pod for the VM
+			temlateService, err := services.NewTemplateService("whatever")
+			Expect(err).ToNot(HaveOccurred())
+			var pod *kubev1.Pod
+			pod, err = temlateService.RenderLaunchManifest(vm)
+			Expect(err).ToNot(HaveOccurred())
+			pod.Spec.NodeName = "mynode"
+
+			// Create the expected VM after the update
+			obj, err := conversion.NewCloner().DeepCopy(vm)
+			Expect(err).ToNot(HaveOccurred())
+
+			expectedVM := obj.(*v1.VM)
+			expectedVM.Status.Phase = v1.Pending
+			expectedVM.Status.NodeName = pod.Spec.NodeName
+			expectedVM.ObjectMeta.Labels = map[string]string{v1.NodeNameLabel: pod.Spec.NodeName}
+
+			// Register the expected REST call
+			server.AppendHandlers(
+				ghttp.CombineHandlers(
+					ghttp.VerifyRequest("PUT", "/apis/kubevirt.io/v1alpha1/namespaces/default/vms/testvm"),
+					ghttp.VerifyJSONRepresenting(expectedVM),
+					ghttp.RespondWithJSONEncoded(http.StatusOK, expectedVM),
+				),
+			)
+
+			// Tell the controller that there is a new running Pod
+
+			queue := workqueue.NewRateLimitingQueue(workqueue.DefaultControllerRateLimiter())
+			key, _ := cache.MetaNamespaceKeyFunc(pod)
+			podCache.Add(pod)
+			queue.Add(key)
+			podDispatch.Execute(podCache, queue, key)
+
+			Expect(len(server.ReceivedRequests())).To(Equal(1))
 			close(done)
 		}, 10)
 	})
